Add Contains/DoesNotContain assert APIs
